feat: booking 엔티티에 bookedAt 컬럼 추가 및 상태 컬럼 추가 / Concert 엔티티에 ticketLimit Integer 래퍼클래스로 변경

+++ src/main/java/io/github/murphscall/concertbooking/concert/domain/Concert.java
@@ -20,7 +20,7 @@ public class Concert extends BaseEntity {
 	private Long id;
 
 	@Column(name = "ticket_limit", nullable = false)
-	private Long ticketLimit;
+	private Integer ticketLimit;
 
 	@Column(name = "name", nullable = false)
 	private String name;
